COMP: Got rid of obsolete code incompatible with ITK4
The itk::RawImageSequenceReder class was a hacked up
version of itk::ImageFileReader -- the author either wrote
it before itk::ImageSeriesReader was available, or was unaware
of it as a way to read an image from a seriesof slice files.

The changes to the Classifier apps had to do with code depending
on ITK_USE_REVIEW_STATISTICS, which is no longer relevant in
ITKV4 as the review statistics framework is now part of ITK proper.

ENH: Use ITK_LIBRARIES in target_link_libraries.

Change-Id: Ic807392a68fd2e0a09f236f68cd3cc2fd0ee2f11