COMP: Update FLTK SVN configuration.
Upstream FLTK changed their SVN server configuration.  The domain name
has changed and a checkout now requires a blank username and password.

Since the currently released version of CMake does not support an empty
username and password in the SVN checkout, a custom DOWNLOAD_COMMAND
is used for the FLTK ExternalProject.

Change-Id: I9f59b84d585bfcb69ad063534e6f21be8f8f165a